pyproject.toml
[tool.poetry]
name = "homework-27"
version = "0.1.0"
description = ""
authors = [" AleshukinRV <rusikk@mai.ru>"]
readme = "README.md"
packages = [{include = "homework_27"}]

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.1.7"
psycopg2 = "^2.9.6"
pillow = "^9.5.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
pytest-django = "^4.5.2"
factory-boy = "^3.2.1"
pytest-factoryboy = "^2.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
